python-discord-bot/pyproject.toml
[project]
name = "thera-vl-discord-bot"
version = "0.1.0"
description = "Discord bot for TheraLab Virtual - a multi-agent AI collaboration system"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "discord.py>=2.3.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "openai>=1.0.0",
    "anthropic>=0.5.0",
    "mistralai>=0.0.7",
    "asyncio>=3.4.3",
    "typing-extensions>=4.7.0",
    "python-dateutil>=2.8.2",
    "psycopg2-binary>=2.9.9",
    "SQLAlchemy>=2.0.0",
    "alembic>=1.12.0",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["commands"]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
]
